+++ src/interface/index.ts
@@ -1,6 +1,4 @@
 export interface Locales<T = any> {
-  /** Chinese */
-  zh_CN: T;
   /** English */
   en_US: T;
 }
