+++ src/pages/layout/header.tsx
@@ -11,7 +11,7 @@ import { ReactComponent as MoonSvg } from '@/assets/header/moon.svg';
 import { ReactComponent as SunSvg } from '@/assets/header/sun.svg';
 import { LocaleFormatter, useLocale } from '@/locales';
 import ReactSvg from '@/assets/logo/react.svg';
-import AntdSvg from '@/assets/logo/antd.svg';
+// import AntdSvg from '@/assets/logo/antd.svg';
 import { logoutAsync, setUserItem } from '@/stores/user.store';
 import { useDispatch, useSelector } from 'react-redux';
 import { setGlobalState } from '@/stores/global.store';
@@ -73,7 +73,7 @@ const HeaderComponent: FC<HeaderProps> = ({ collapsed, toggle }) => {
       {device !== 'MOBILE' && (
         <div className="logo" style={{ width: collapsed ? 80 : 200 }}>
           <img src={ReactSvg} alt="" style={{ marginRight: collapsed ? '2px' : '20px' }} />
-          <img src={AntdSvg} alt="" />
+          {/* <img src={AntdSvg} alt="" /> */}
         </div>
       )}
       <div className="layout-page-header-main">
